\name{brain}
\docType{data}
\alias{brain}
\title{7 brain cancer studies}
\description{
7 brain cancer studies comparing Anaplastic Astrocytoma (AA) and Glioblastoma multiforme (GBM) samples.
\tabular{lllll}{
Data Name \tab Published Year   \tab Array Platform \tab Sample Size \tab GEO Accession ID\cr
Freije \tab 2004   \tab HG-U133A,B \tab 85 \tab GSE4412\cr
Phillips \tab 2006 \tab HG-U133A,B \tab 100 \tab GSE4271\cr
Sun \tab 2006 \tab HG-U133 Plus 2 \tab 100 \tab GSE4290\cr
Yamanaka \tab 2006 \tab Agilent \tab 29 \tab GSE4381\cr
Petalidis \tab 2008 \tab HG-U133A \tab 58 \tab GSE1993\cr
Gravendeel \tab 2009 \tab HG-U133 Plus 2 \tab 175 \tab GSE16011\cr
Sun \tab 2010 \tab HG-U133 Plus 2 \tab 42 \tab GSE19578\cr
}
}
\usage{brain}
\format{A list containing 7 matrices. Each matrix is gene expression data after gene filtering.}
\source{Gene Expression Omnibus (GEO)}
\references{
Freije W, Castro-Vargas F, Fang Z, Horvath S, Cloughesy T, et al. (2004) Gene expression profiling of gliomas strongly predicts survival. Cancer research 64: 6503.  

Phillips H, Kharbanda S, Chen R, Forrest W, Soriano R, et al. (2006) Molecular subclasses of high-grade glioma predict prognosis, delineate a pattern of disease progression, and resemble stages in neurogenesis. Cancer Cell 9: 157-173.  

Sun L, Hui A, Su Q, Vortmeyer A, Kotliarov Y, et al. (2006) Neuronal and glioma-derived stem cell factor induces angiogenesis within the brain. Cancer cell 9: 287-300.  

Yamanaka R, Arao T, Yajima N, Tsuchiya N, Homma J, et al. (2006) Identification of expressed genes characterizing long-term survival in malignant glioma patients. Oncogene 25: 5994-6002.  

Petalidis L, Oulas A, Backlund M, Wayland M, Liu L, et al. (2008) Improved grading and survival prediction of human astrocytic brain tumors by artificial neural network analysis of gene expression microarray data. Molecular cancer therapeutics 7: 1013.  

Gravendeel L, Kouwenhoven M, Gevaert O, de Rooi J, Stubbs A, et al. (2009) Intrinsic gene expression profiles of gliomas are a better predictor of survival than histology. Cancer research 69: 9065.  

Paugh B, Qu C, Jones C, Liu Z, Adamowicz-Brice M, et al. (2010) Integrated molecular genetic profiling of pediatric high-grade gliomas reveals key differences with the adult disease. Journal of Clinical Oncology 28: 3061.  

}
\keyword{MetaQC}
